Start to refactor the layout sharing screen into functional components in their own files
