--- src/renderer/screens/Editor/Sidebar/Overview/LayoutSharing.js
@@ -1,501 +1,0 @@
-// -*- mode: js-jsx -*-
-/* Chrysalis -- Kaleidoscope Command Center
- * Copyright (C) 2020-2022  Keyboardio, Inc.
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, version 3.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import Focus from "@api/focus";
-import { KeymapDB } from "@api/keymap";
-import Log from "@api/log";
-import CloseIcon from "@mui/icons-material/Close";
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
-import Dialog from "@mui/material/Dialog";
-import DialogContent from "@mui/material/DialogContent";
-import DialogTitle from "@mui/material/DialogTitle";
-import Divider from "@mui/material/Divider";
-import Drawer from "@mui/material/Drawer";
-import IconButton from "@mui/material/IconButton";
-import MenuItem from "@mui/material/MenuItem";
-import MenuList from "@mui/material/MenuList";
-import Typography from "@mui/material/Typography";
-import withStyles from "@mui/styles/withStyles";
-import ConfirmationDialog from "@renderer/components/ConfirmationDialog";
-import { getStaticPath } from "@renderer/config";
-import { ipcRenderer } from "electron";
-import fs from "fs";
-import i18n from "i18next";
-import jsonStringify from "json-stringify-pretty-compact";
-import path from "path";
-import React from "react";
-import { useTranslation } from "react-i18next";
-import { toast } from "react-toastify";
-const db = new KeymapDB();
-const focus = new Focus();
-
-const sidebarWidth = 300;
-
-const styles = (theme) => ({
-  closeButton: {
-    position: "absolute",
-    right: theme.spacing(1),
-    top: theme.spacing(1),
-    color: theme.palette.grey[500],
-  },
-  toolbar: {
-    display: "flex",
-    flexWrap: "wrap",
-    marginBottom: theme.spacing(2),
-    marginTop: -theme.spacing(1),
-  },
-  libraryImport: {
-    minWidth: "15em",
-    marginRight: theme.spacing(2),
-    marginBottom: theme.spacing(2),
-  },
-  libraryImportRoot: {
-    marginBottom: theme.spacing(2),
-  },
-  fileImportRoot: {
-    marginBottom: theme.spacing(2),
-  },
-  fileExportRoot: {
-    marginBottom: theme.spacing(2),
-  },
-  drawer: {
-    width: sidebarWidth,
-    flexShrink: 0,
-  },
-  drawerPaper: {
-    width: sidebarWidth,
-    marginTop: 65,
-  },
-  drawerContainer: {
-    overflow: "auto",
-    padding: theme.spacing(3),
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3),
-    marginLeft: sidebarWidth,
-    width: `calc(100% - ${sidebarWidth}px)`,
-  },
-});
-
-const loadLayout = (fileName, fileData) => {
-  const logger = new Log();
-
-  if (!fileData) {
-    try {
-      fileData = fs.readFileSync(fileName);
-    } catch (e) {
-      logger.error("Unable to read layout", {
-        filename: fileName,
-        error: e.message,
-      });
-      toast.error(i18n.t("editor.sharing.errors.unableToLoad"));
-      return null;
-    }
-  }
-
-  if (!fileData) {
-    logger.error("Unable to read layout", {
-      filename: fileName,
-    });
-
-    toast.error(i18n.t("editor.sharing.errors.unableToLoad"));
-    return null;
-  }
-
-  let layoutData;
-  try {
-    layoutData = JSON.parse(new TextDecoder().decode(fileData));
-  } catch (e) {
-    logger.error("Failed to parse layout JSON", {
-      filename: fileName,
-      error: e.message,
-    });
-    toast.error(i18n.t("editor.sharing.errors.parseFail"));
-    return null;
-  }
-
-  const src_keymaps = layoutData.keymaps || layoutData.keymap.custom;
-
-  let keymaps;
-  try {
-    keymaps = src_keymaps.map((layer) => {
-      return layer.map((key) => {
-        return db.lookup(key.keyCode || key.code);
-      });
-    });
-  } catch (_) {
-    logger.error("Layout file did not contain valid layout data", {
-      filename: fileName,
-    });
-    toast.error(i18n.t("editor.sharing.errors.invalidLayoutData"));
-    return null;
-  }
-
-  return {
-    keymaps: keymaps,
-    colormaps: layoutData.colormaps || layoutData.colormap?.colorMap,
-    palette: layoutData.palette || layoutData.colormap?.palette,
-  };
-};
-
-const LibraryImport = (props) => {
-  const selectLibraryItem = (item) => () => {
-    loadFromLibrary(item);
-  };
-
-  const loadFromLibrary = (layoutName) => {
-    const { vendor, product } = focus.focusDeviceDescriptor.info;
-    const cVendor = vendor.replace("/", "");
-    const cProduct = product.replace("/", "");
-    const layoutPath = (layout) =>
-      path.join(getStaticPath(), cVendor, cProduct, `layouts/${layout}.json`);
-
-    const layoutData = loadLayout(layoutPath(layoutName));
-
-    if (layoutData != null) props.setLayout(layoutName, layoutData);
-  };
-
-  const { classes, library, layoutName } = props;
-
-  if (library.length == 0) return null;
-
-  const layouts = library.map((name) => {
-    const label = name.charAt(0).toUpperCase() + name.slice(1);
-
-    return (
-      <MenuItem
-        selected={layoutName == name}
-        value={name}
-        key={`library-item-${name}`}
-        onClick={selectLibraryItem(name)}
-      >
-        {label}
-      </MenuItem>
-    );
-  });
-
-  return (
-    <Box sx={{ sb: 2 }}>
-      <Typography variant="h5">
-        {i18n.t("editor.sharing.loadFromLibrary")}
-      </Typography>
-      <MenuList>{layouts}</MenuList>
-
-      <Divider />
-    </Box>
-  );
-};
-
-class BackupImportBase extends React.Component {
-  state = { library: [] };
-
-  selectBackupItem = (item) => () => {
-    const [layoutFileData, error] = ipcRenderer.sendSync(
-      "backups.load-file",
-      focus.focusDeviceDescriptor.info,
-      item
-    );
-
-    if (error) {
-      // TODO(anyone): show toast
-      console.error(error);
-      return;
-    }
-
-    const layoutData = loadLayout("$userData/" + item, layoutFileData);
-    if (layoutData != null) this.props.setLayout(item, layoutData);
-  };
-
-  componentDidMount() {
-    const library = ipcRenderer.sendSync(
-      "backups.list-library",
-      focus.focusDeviceDescriptor.info
-    );
-
-    this.setState({ library: library });
-  }
-
-  formatName = (name) => {
-    const ts = new Date(parseInt(name));
-
-    return ts.toISOString();
-  };
-
-  render() {
-    const { classes, layoutName } = this.props;
-    const { library } = this.state;
-
-    if (library.length == 0) return null;
-
-    const layouts = library.map((name) => {
-      const label = this.formatName(name);
-
-      return (
-        <MenuItem
-          selected={layoutName == name}
-          value={name}
-          key={`backup-item-${name}`}
-          onClick={this.selectBackupItem(name)}
-        >
-          {label}
-        </MenuItem>
-      );
-    });
-
-    return (
-      <Box sx={{ sb: 2 }}>
-        <Typography variant="h5">
-          {i18n.t("editor.sharing.loadFromBackup")}
-        </Typography>
-        <MenuList>{layouts}</MenuList>
-
-        <Divider />
-      </Box>
-    );
-  }
-}
-
-const BackupImport = withStyles(styles, { withTheme: true })(BackupImportBase);
-
-const FileImport = (props) => {
-  const importFromFile = () => {
-    const [fileName, fileData] = ipcRenderer.sendSync("file-open", {
-      title: i18n.t("editor.sharing.selectLoadFile"),
-      filters: [
-        {
-          name: i18n.t("editor.sharing.dialog.layoutFiles"),
-          extensions: ["json", "layout"],
-        },
-        {
-          name: i18n.t("editor.sharing.dialog.allFiles"),
-          extensions: ["*"],
-        },
-      ],
-    });
-    // If we have no filename, then the dialog was canceled.
-    if (!fileName) return;
-    const layoutData = loadLayout(fileName, fileData);
-    if (layoutData != null) props.setLayout("custom", layoutData);
-  };
-
-  return (
-    <Box sx={{ mb: 2 }}>
-      <Button variant="outlined" onClick={importFromFile}>
-        {i18n.t("editor.sharing.loadFromFile")}
-      </Button>
-    </Box>
-  );
-};
-
-const ExportToFileBase = (props) => {
-  const { t } = useTranslation();
-  const exportToFile = () => {
-    const { keymap, colormap } = props;
-    const data = {
-      keymaps: keymap.custom,
-      colormaps: colormap.colorMap,
-      palette: colormap.palette,
-    };
-    ipcRenderer.send("file-save", {
-      content: jsonStringify(data),
-      title: i18n.t("editor.sharing.selectExportFile"),
-      defaultPath: "Layout.json",
-      filters: [
-        {
-          name: i18n.t("editor.sharing.dialog.layoutFiles"),
-          extensions: ["json", "layout"],
-        },
-        {
-          name: i18n.t("editor.sharing.dialog.allFiles"),
-          extensions: ["*"],
-        },
-      ],
-    });
-  };
-  return (
-    <Box sx={{ mb: 2 }}>
-      <Button variant="outlined" onClick={exportToFile}>
-        {i18n.t("editor.sharing.exportToFile")}
-      </Button>
-    </Box>
-  );
-};
-const ExportToFile = withStyles(styles, { withTheme: true })(ExportToFileBase);
-
-class LayoutSharingBase extends React.Component {
-  constructor(props) {
-    super(props);
-
-    const { vendor, product } = focus.focusDeviceDescriptor.info;
-    const cVendor = vendor.replace("/", "");
-    const cProduct = product.replace("/", "");
-    const layoutDirPath = path.join(
-      getStaticPath(),
-      cVendor,
-      cProduct,
-      "layouts"
-    );
-
-    try {
-      const layouts = fs
-        .readdirSync(layoutDirPath, {
-          encoding: "utf-8",
-        })
-        .map((name) => path.basename(name, ".json"))
-        .sort();
-
-      this.state = {
-        importConfirmOpen: false,
-        layout: {},
-        layoutName: null,
-        library: layouts,
-      };
-    } catch (_) {
-      this.state = {
-        importConfirmOpen: false,
-        layout: {},
-        layoutName: null,
-        library: [],
-      };
-    }
-  }
-
-  setLayout = (layoutName, layout) => {
-    this.setState({
-      layoutName: layoutName,
-      layout: layout,
-    });
-  };
-
-  openImportConfirm = () => {
-    this.setState({ importConfirmOpen: true });
-  };
-  closeImportConfirm = () => {
-    this.setState({ importConfirmOpen: false });
-  };
-
-  onImport = () => {
-    const { keymap, colormap } = this.props;
-    const { layout } = this.state;
-
-    const newKeymap = layout.keymaps.concat(
-      keymap.custom.slice(layout.keymaps.length)
-    );
-    const newColormap = layout.colormaps
-      ? layout.colormaps.concat(
-          colormap.colorMap.slice(layout.colormaps.length)
-        )
-      : colormap.colorMap;
-    const newPalette = layout.palette || colormap.palette;
-
-    this.closeImportConfirm();
-    this.props.onKeymapChange(newKeymap);
-    this.props.onColormapChange(newColormap);
-    this.props.onPaletteChange(newPalette);
-    this.props.onClose();
-  };
-
-  render() {
-    const { classes, open, onClose, theme, keymap, colormap, ...others } =
-      this.props;
-    const { layout, layoutName, library } = this.state;
-
-    const Keymap = focus.focusDeviceDescriptor.components.keymap;
-    const previewLayout = layout.keymaps ? layout.keymaps[0] : keymap.custom[0];
-    const palette = layout.palette || colormap.palette;
-    const previewColormap = layout.colormaps
-      ? layout.colormaps[0]
-      : colormap.colorMap[0];
-    const keymapWidget = (
-      <Keymap
-        className="layer"
-        keymap={previewLayout}
-        palette={palette}
-        colormap={previewColormap}
-        theme={theme}
-      />
-    );
-
-    return (
-      <Dialog open={open} onClose={onClose} fullScreen>
-        <DialogTitle>
-          <Typography>{i18n.t("editor.sharing.title")}</Typography>
-          <IconButton
-            onClick={onClose}
-            className={classes.closeButton}
-            size="large"
-          >
-            <CloseIcon />
-          </IconButton>
-        </DialogTitle>
-        <Divider />
-        <DialogContent>
-          <Drawer
-            className={classes.drawer}
-            variant="permanent"
-            anchor="left"
-            classes={{
-              paper: classes.drawerPaper,
-            }}
-          >
-            <div className={classes.drawerContainer}>
-              <LibraryImport
-                setLayout={this.setLayout}
-                library={library}
-                layoutName={layoutName}
-                {...others}
-              />
-              <BackupImport
-                setLayout={this.setLayout}
-                layoutName={layoutName}
-                {...others}
-              />
-              <FileImport setLayout={this.setLayout} {...others} />
-              <ExportToFile keymap={keymap} colormap={colormap} />
-
-              <Button
-                disabled={layoutName == null}
-                variant="outlined"
-                color="primary"
-                onClick={this.openImportConfirm}
-              >
-                {i18n.t("editor.sharing.import")}
-              </Button>
-              <ConfirmationDialog
-                title={i18n.t("editor.sharing.importConfirm.title")}
-                open={this.state.importConfirmOpen}
-                onConfirm={this.onImport}
-                onCancel={this.closeImportConfirm}
-              >
-                {i18n.t("editor.sharing.importConfirm.contents")}
-              </ConfirmationDialog>
-            </div>
-          </Drawer>
-
-          <div className={classes.content}>{keymapWidget}</div>
-        </DialogContent>
-      </Dialog>
-    );
-  }
-}
-const LayoutSharing = withStyles(styles, { withTheme: true })(
-  LayoutSharingBase
-);
-
-export { LayoutSharing as default };
